Time: 0 ms (100%), Space: 13.8 MB (48.06%) - LeetHub

+++ 1822-sign-of-the-product-of-an-array/1822-sign-of-the-product-of-an-array.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int arraySign(vector<int>& nums) {
-        long long x = 1;
+        int x = 1;
 
         for(int i=0; i<nums.size(); i++){
             if(nums[i] > 0){
